Database is up and running in the project, basic chat functionalities implemented (ChatActivity, renamed from ChatroomActivity, MessageAdapter and Message Java classes)
Restructuring the way the predefined chatrooms are set up, now they have their own object type (Chatroom.java). The reason for the change is to make the clode more clean and to make it easier to implement the refresh functionality in the lobby. Updated ChatroomAdapter class accordingly, still working on ChatLobbyActivity.

Finalized a layout XML (message_layout.xml), the layout for a single message element in the ChatActivity.